+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.2.6] / 2023-03-28
+## [1.2.6] / 2023-04-11
 ### Features
 - [x] Tests with abstract class works like a [Test] methods.
 - [x] Tests alias with arguments show type parameters.